Rename module decoder to evm_decoder (29-diamond-event)
