+++ proxy/indexer/indexer.py
@@ -17,7 +17,7 @@ from ..indexer.canceller import Canceller
 from ..common_neon.utils import NeonTxResultInfo, NeonTxInfo, str_fmt_object
 from ..common_neon.solana_interactor import SolanaInteractor
 from ..common_neon.solana_receipt_parser import SolReceiptParser
-from ..common_neon.decoder import decode
+from ..common_neon.evm_decoder import decode
 from ..common_neon.environment_data import EVM_LOADER_ID, FINALIZED, CANCEL_TIMEOUT, SKIP_CANCEL_TIMEOUT, HOLDER_TIMEOUT
